feat: remove link on remove format if selection is a link
