+++ src/Buttons.js
@@ -7,6 +7,7 @@ import React, {
 } from "react";
 import { EditorContext } from "./EditorContext";
 import { uploadImages } from "./ImageUpload/utils";
+import { isLinkNode, removeLink } from "./link/utils";
 
 export function MarkButton({ type, children, value, onMouseDown, ...rest }) {
   const editor = useContext(EditorContext);
@@ -147,6 +148,11 @@ export function RemoveFormatButton({ ...rest }) {
           return;
         }
         editor.formatter.toggle("removeformat");
+        const node = editor.selection.getNode();
+
+        if (isLinkNode(editor, node)) {
+          removeLink(editor);
+        }
       }}
     >
       Remove format
